pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hw-router"
version = "0.1.0"
description = "HW-Router: Hardware-Aware Routing for Scalable Multi-LLM Serving (DAC 2026)"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    { name = "Ahasan Kabir", email = "ahasankabir146@gmail.com" }
]

dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "requests>=2.28",
    "tqdm>=4.60",
]

[project.optional-dependencies]
irt = [
    "transformers>=4.36",
]
carrot = [
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
]
umr = [
    "sentence-transformers>=2.2",
]
serving = [
    "vllm>=0.3",
    "openai>=1.0",
    "fastapi>=0.100",
    "uvicorn>=0.24",
    "httpx>=0.25",
    "prometheus-client>=0.17",
]
viz = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
]
dev = [
    "ruff>=0.1",
]
all = [
    "hw-router[irt,carrot,umr,serving,viz,dev]",
]

[project.urls]
Repository = "https://github.com/Ahasannn/hardware-aware-router"

[tool.setuptools.packages.find]
include = ["hw_router*", "baselines*"]

[tool.ruff]
line-length = 100
target-version = "py39"
